+++ src/android/sdk/oracle/idm/mobile/auth/local/SecretKeyWrapper.java
@@ -106,6 +106,7 @@ public class SecretKeyWrapper {
 
     /**
      * Generate key pair.
+     *
      * @param context
      * @param alias
      * @throws GeneralSecurityException
@@ -119,10 +120,13 @@ public class SecretKeyWrapper {
                     .setEncryptionPaddings(KeyProperties.ENCRYPTION_PADDING_RSA_PKCS1)
                     .setUserAuthenticationRequired(true)
                     .build();
-            //TODO Update studio and build to Android N, then set jdk location in File->Project structure
-//            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-//                builder.setInvalidatedByBiometricEnrollment(false);
-//            }
+            /*KeyGenParameterSpec.Builder#setInvalidatedByBiometricEnrollment(true)
+            * By default, this is true, so these keys are irreversibly invalidated when a new
+            * fingerprint is enrolled, or when all existing fingerprints are deleted.
+            * This is considered more secure. For more details, refer javadoc
+            * of setInvalidatedByBiometricEnrollment.
+            * */
+
             final KeyPairGenerator gen = KeyPairGenerator.getInstance(KeyProperties.KEY_ALGORITHM_RSA, "AndroidKeyStore");
             gen.initialize(spec);
             gen.generateKeyPair();